Distinguished between L of gas and L of solution to reduce errors caused by dropdown menus.
This change has been implemented only for the web interface, and seems
quite hacky, but helpful for new users.

Clarified the help messages: simpler language, better spacing.
I don't want to make L (g) vs L (sol) mandatory for the CLI because it's
annoying.

However, this should still be an option for programs built ontop of the
chemcli DSL.